requirements.txt
svgwrite-1.4.3